+++ src/DevAdventCalendarCompetition/DevAdventCalendarCompetition.Tests/AutoFixture/TestWithNumberWithoutAnswersDataAttribute.cs
@@ -6,17 +6,17 @@ using AutoFixture.Xunit2;
 namespace DevAdventCalendarCompetition.Tests.AutoFixture
 {
     [AttributeUsage(AttributeTargets.Method)]
-    public sealed class AutoDataWithoutChildsAttribute : AutoDataAttribute
+    public sealed class TestWithNumberWithoutAnswersDataAttribute : AutoDataAttribute
     {
-        public AutoDataWithoutChildsAttribute()
+        public TestWithNumberWithoutAnswersDataAttribute()
             : base(() =>
             {
                 var fixture = new Fixture();
                 fixture.Behaviors.OfType<ThrowingRecursionBehavior>().ToList()
                     .ForEach(b => fixture.Behaviors.Remove(b));
                 fixture.Behaviors.Add(new OmitOnRecursionBehavior());
-                fixture.Customizations.Add(new TestNumberGenerator());
-                fixture.Customizations.Add(new TestWithoutChildsGenerator());
+                fixture.Customizations.Add(new TestNumberSpecimen());
+                fixture.Customizations.Add(new TestWithoutAnswersSpecimen());
                 return fixture;
             })
         {
